# src/main/java/pl/wsb/students/api/model/MovieList.java
/*
 * Java WSB Course Sample Movie App API
 * Sample Movie App API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package pl.wsb.students.api.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.annotations.ApiModelProperty;

import javax.validation.constraints.NotNull;
import java.util.ArrayList;
import java.util.List;
import java.util.Objects;

/**
 * MovieList
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2020-04-03T15:00:27.698Z")
public class MovieList   {
  @JsonProperty("total")
  private Integer total = null;

  @JsonProperty("data")
  private List<Movie> data = new ArrayList<Movie>();

  public MovieList total(Integer total) {
    this.total = total;
    return this;
  }

  /**
   * Movie list total items count
   * @return total
   **/
  @JsonProperty("total")
  @ApiModelProperty(required = true, value = "Movie list total items count")
  @NotNull
  public Integer getTotal() {
    return total;
  }

  public void setTotal(Integer total) {
    this.total = total;
  }

  public MovieList data(List<Movie> data) {
    this.data = data;
    return this;
  }

  public MovieList addDataItem(Movie dataItem) {
    this.data.add(dataItem);
    return this;
  }

  /**
   * One pagination page data
   * @return data
   **/
  @JsonProperty("data")
  @ApiModelProperty(required = true, value = "One pagination page data")
  @NotNull
  public List<Movie> getData() {
    return data;
  }

  public void setData(List<Movie> data) {
    this.data = data;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    MovieList movieList = (MovieList) o;
    return Objects.equals(this.total, movieList.total) &&
        Objects.equals(this.data, movieList.data);
  }

  @Override
  public int hashCode() {
    return Objects.hash(total, data);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class MovieList {\n");
    
    sb.append("    total: ").append(toIndentedString(total)).append("\n");
    sb.append("    data: ").append(toIndentedString(data)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}